Remove vadmin stuff, name parser command 'vparse'
