--- src/vivarium_cluster_tools/shared.py
@@ -1,26 +1,0 @@
-from bdb import BdbQuit
-import functools
-
-from loguru import logger
-
-
-def handle_exceptions(func, with_debugger):
-    """Drops a user into an interactive debugger if func raises an error."""
-
-    @functools.wraps(func)
-    def wrapped(*args, **kwargs):
-        try:
-            func(*args, **kwargs)
-        except (BdbQuit, KeyboardInterrupt):
-            raise
-        except Exception as e:
-            logger.exception("Uncaught exception {}".format(e))
-            if with_debugger:
-                import pdb
-                import traceback
-                traceback.print_exc()
-                pdb.post_mortem()
-            else:
-                raise
-
-    return wrapped
